Move unit tests to a separate directory and update
