+++ tests/Unit/GeneratorsProviderTest.php
@@ -1,12 +1,12 @@
 <?php
 
-namespace Fakturoid\Tests;
+namespace Fakturoid\Tests\Unit;
 
 use Fakturoid\Dispatcher;
-use Fakturoid\Provider\InventoryMovesProvider;
+use Fakturoid\Provider\GeneratorsProvider;
 use Fakturoid\Response;
 
-class InventoryMovesProviderTest extends \Fakturoid\Tests\TestCase
+class GeneratorsProviderTest extends UnitTestCase
 {
     public function testList(): void
     {
@@ -15,10 +15,10 @@ class InventoryMovesProviderTest extends \Fakturoid\Tests\TestCase
 
         $dispatcher->expects($this->once())
             ->method('get')
-            ->with('/accounts/{accountSlug}/inventory_moves.json', ['page' => 1])
+            ->with('/accounts/{accountSlug}/generators.json', ['page' => 1])
             ->willReturn(new Response($responseInterface));
 
-        $provider = new InventoryMovesProvider($dispatcher);
+        $provider = new GeneratorsProvider($dispatcher);
         $response = $provider->list(['page' => 1]);
         $this->assertEquals([], $response->getBody(true));
     }
@@ -26,34 +26,32 @@ class InventoryMovesProviderTest extends \Fakturoid\Tests\TestCase
     public function testGet(): void
     {
         $dispatcher = $this->createMock(Dispatcher::class);
-        $responseInterface = $this->createPsrResponseMock(200, 'application/json', '{}');
 
-        $itemId = 8;
-        $moveId = 60;
+        $id = 6;
+        $responseInterface = $this->createPsrResponseMock(200, 'application/json', '{"page": 2}');
         $dispatcher->expects($this->once())
             ->method('get')
-            ->with(sprintf('/accounts/{accountSlug}/inventory_items/%d/inventory_moves/%d.json', $itemId, $moveId))
+            ->with(sprintf('/accounts/{accountSlug}/generators/%d.json', $id))
             ->willReturn(new Response($responseInterface));
 
-        $provider = new InventoryMovesProvider($dispatcher);
-        $response = $provider->get($itemId, $moveId);
-        $this->assertEquals([], $response->getBody(true));
+        $provider = new GeneratorsProvider($dispatcher);
+        $response = $provider->get($id);
+        $this->assertEquals(['page' => 2], $response->getBody(true));
     }
 
     public function testDelete(): void
     {
         $dispatcher = $this->createMock(Dispatcher::class);
 
-        $itemId = 8;
-        $moveId = 60;
+        $id = 6;
         $responseInterface = $this->createPsrResponseMock(200, 'application/json', '{"page": 2}');
         $dispatcher->expects($this->once())
             ->method('delete')
-            ->with(sprintf('/accounts/{accountSlug}/inventory_items/%d/inventory_moves/%d.json', $itemId, $moveId))
+            ->with(sprintf('/accounts/{accountSlug}/generators/%d.json', $id))
             ->willReturn(new Response($responseInterface));
 
-        $provider = new InventoryMovesProvider($dispatcher);
-        $response = $provider->delete($itemId, $moveId);
+        $provider = new GeneratorsProvider($dispatcher);
+        $response = $provider->delete($id);
         $this->assertEquals(['page' => 2], $response->getBody(true));
     }
 
@@ -61,16 +59,15 @@ class InventoryMovesProviderTest extends \Fakturoid\Tests\TestCase
     {
         $dispatcher = $this->createMock(Dispatcher::class);
 
-        $itemId = 8;
-        $moveId = 60;
+        $id = 6;
         $responseInterface = $this->createPsrResponseMock(200, 'application/json', '{"page": 2}');
         $dispatcher->expects($this->once())
             ->method('patch')
-            ->with(sprintf('/accounts/{accountSlug}/inventory_items/%d/inventory_moves/%d.json', $itemId, $moveId))
+            ->with(sprintf('/accounts/{accountSlug}/generators/%d.json', $id))
             ->willReturn(new Response($responseInterface));
 
-        $provider = new InventoryMovesProvider($dispatcher);
-        $response = $provider->update($itemId, $moveId, ['page' => 2]);
+        $provider = new GeneratorsProvider($dispatcher);
+        $response = $provider->update($id, ['page' => 2]);
         $this->assertEquals(['page' => 2], $response->getBody(true));
     }
 
@@ -78,15 +75,14 @@ class InventoryMovesProviderTest extends \Fakturoid\Tests\TestCase
     {
         $dispatcher = $this->createMock(Dispatcher::class);
 
-        $itemId = 8;
         $responseInterface = $this->createPsrResponseMock(200, 'application/json', '{"page": 2}');
         $dispatcher->expects($this->once())
             ->method('post')
-            ->with(sprintf('/accounts/{accountSlug}/inventory_items/%d/inventory_moves.json', $itemId))
+            ->with('/accounts/{accountSlug}/generators.json')
             ->willReturn(new Response($responseInterface));
 
-        $provider = new InventoryMovesProvider($dispatcher);
-        $response = $provider->create($itemId, ['page' => 2]);
+        $provider = new GeneratorsProvider($dispatcher);
+        $response = $provider->create(['page' => 2]);
         $this->assertEquals(['page' => 2], $response->getBody(true));
     }
 }
